perf(v312-59-D): add jemalloc-prof feature + heap dump via SIGUSR2
Build a profiling-enabled jemalloc as the global allocator (off by
default; opt-in with `--features jemalloc-prof`). The profiling
hot path is a SIGUSR2 handler that drops `jeprof.<pid>.<ts>.heap`
into /tmp without restarting the server, so SOAK runs can take
periodic snapshots and the next session can diff them with
`jeprof --text --base=<baseline> <binary> <snapshot>`.

Findings while wiring this up (relevant for SOAK 5691 follow-up):

- tikv-jemalloc-sys 0.7.1 always passes `--with-malloc-conf` to
  jemalloc's configure (defaults to `background_thread:...` on
  macOS). Once `--with-malloc-conf` is set the runtime `MALLOC_CONF`
  env var is **silently ignored**. The Cargo.toml comment now spells
  out the required
  `JEMALLOC_SYS_WITH_MALLOC_CONF="prof:true,prof_active:true,lg_prof_sample:14"`
  build env var; without it `opt.prof` stays false even if
  `MALLOC_CONF="prof:true"` is exported at runtime (verified:
  `mallctl("opt.prof")` returns false, `prof.dump` returns ENOENT).
- jemalloc 5.3.x's `prof.dump` WRITE-macro expects
  `newlen == sizeof(const char *)` (8 bytes on 64-bit), not the
  string length. Passing the raw path bytes triggers EINVAL; passing
  `&path_cstr + sizeof(*)` makes `prof.dump` return 0 and write the
  heap file. The handler now does this correctly.
- Smoke test (300 inserts + 300 selects, 4 SIGUSR2 snapshots)
  confirms `opt.prof=true`, `opt.lg_prof_sample=14`, `prof.dump r=0`
  and produces four ~37KB `.heap` files. `jeprof --text --base=…`
  diff shows the expected per-txn growth
  (`commit_implicit_dml_tx` / `execute_insert` / `commit_transaction` /
  `clone` / `append_elements` — all bounded per row) and total RSS
  stays under 26MB across the run.

Also includes two small follow-ups from PR #4522 / #4523:

- `crates/executor/src/expr/mod.rs`: drop redundant `as i64` cast
  in `strftime_value` (clippy `unnecessary_cast`).
- `crates/storage/src/file_storage.rs` and
  `crates/transaction/src/transaction_manager.rs`: rustfmt fixes.

Issue: V312-59-D / SOAK 5691 leak hunt.

+++ crates/executor/src/expr/mod.rs
@@ -2482,7 +2482,7 @@ fn strftime_value(args: &[Value]) -> Value {
 
     // Weekday (0=Sunday..6=Saturday). days_from_civil(1970,1,1) was
     // Thursday (weekday=4). Adjust by hand from a known anchor.
-    let weekday = ((doy + 3).rem_euclid(7)) as i64;
+    let weekday = (doy + 3).rem_euclid(7);
 
     let mut out = String::with_capacity(fmt.len() + 8);
     let bytes = fmt.as_bytes();
